requirements.txt

gradio==4.44.0
tensorflow-cpu==2.15.0
numpy
pillow
huggingface_hub
